exams/grade7b_math/requirements.txt
python-docx>=1.1.2
reportlab>=4.2.5
Pillow>=11.0.0









